senko-app/project/migrations/manage.py
#!/usr/bin/env python
from migrate.versioning.shell import main
import os

if __name__ == '__main__':
    main(debug='False', url='mysql+pymysql://{user}:{password}@{host}/{database}?charset=utf8'.format(**{
        'user': os.getenv('DB_USER', 'root'),
        'password': os.getenv('DB_PASSWORD', 'password'),
        'host': os.getenv('DB_HOST', 'localhost'),
        'database': os.getenv('DB_DATABASE', 'senko'),
    }), repository='.')
